requirements.txt
antlr4-python3-runtime==4.7.2
asgiref==3.2.5
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.8.2
bech32==1.2.0
bidict==0.19.0
bip32utils==0.3.post4
bleach==3.1.3
bokeh==1.4.0
bs4==0.0.1
bytecode==0.11.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
dask==2.9.2
dataprep==0.2.0
decorator==4.4.2
defusedxml==0.6.0
distributed==2.12.0
Django==3.0.4
djangorestframework==3.11.0
download-youtube-subtitle==0.0.8
drf-yasg==1.17.1
ecdsa==0.13.3
entrypoints==0.3
fastjsonschema==2.14.4
fire==0.2.1
flake8==3.7.9
flake8-click==0.3.1
flake8-requests==0.4.0
Flask==1.1.1
flex==6.14.1
fsspec==0.6.3
hannakageul==0.1.2
HeapDict==1.0.1
holoviews==1.12.7
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.16.0
jigu==0.1.2
Jinja2==2.10.3
jsonpath2==0.4.3
jsonpointer==2.0
jsonschema==3.1.1
jupyter-core==4.6.3
kiwisolver==1.1.0
libcst==0.2.7
listoapi==0.1.47
locket==0.2.0
lxml==4.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mnemonic==0.18
msgpack==1.0.0
mypy-extensions==0.4.3
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
numpy==1.17.5
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
param==1.9.3
parso==0.6.2
partd==1.1.0
pathspec==0.7.0
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
probscale==0.2.3
prompt-toolkit==3.0.4
psutil==5.7.0
ptyprocess==0.6.0
py-geohash-any==1.1
pybembel==0.2.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-box==4.0.4
python-dateutil==2.8.1
pytz==2019.3
pyviz-comms==0.7.4
PyYAML==5.3.1
r2c-py-ast==0.1.0b1
requests==2.22.0
requests-futures==1.0.0
retype==19.9.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.3.3
selenium==3.141.0
setuptools-odoo==2.5.5
setuptools-scm==3.5.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0
sqlparse==0.3.1
strict-rfc3339==0.7
tabulate==0.8.6
tblib==1.6.0
termcolor==1.1.0
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
typing-inspect==0.5.0
uritemplate==3.0.1
urllib3==1.25.8
validate-email==1.3
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.0
wrapt==1.11.2
zict==2.0.0
zipp==3.1.0
